+++ lib/page/login/state/fluent.dart
@@ -1,56 +1,13 @@
-/*
- * Copyright (C) 2020. by perol_notsf, All rights reserved
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program. If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-import 'dart:io';
-
 import 'package:bot_toast/bot_toast.dart';
 import 'package:fluent_ui/fluent_ui.dart';
-import 'package:flutter/widgets.dart';
-import 'package:pixez/custom_tab_plugin.dart';
-import 'package:pixez/er/leader.dart';
 import 'package:pixez/i18n.dart';
-import 'package:pixez/main.dart';
 import 'package:pixez/network/oauth_client.dart';
 import 'package:pixez/page/about/about_page.dart';
 import 'package:pixez/page/hello/setting/setting_quality_page.dart';
-import 'package:pixez/page/webview/webview_page.dart';
-import 'package:pixez/weiss_plugin.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:pixez/page/login/login_page.dart';
+import 'package:url_launcher/url_launcher.dart' as url_launcher;
 
-class FluentLoginPage extends StatefulWidget {
-  @override
-  _FluentLoginPageState createState() => _FluentLoginPageState();
-}
-
-class _FluentLoginPageState extends State<FluentLoginPage> {
-  TextEditingController userNameController = TextEditingController();
-  TextEditingController passWordController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    userNameController.dispose();
-    passWordController.dispose();
-    super.dispose();
-  }
-
+class FluentLoginPageState extends LoginPageStateBase {
   @override
   Widget build(BuildContext context) {
     return _buildBody(context);
@@ -64,7 +21,7 @@ class _FluentLoginPageState extends State<FluentLoginPage> {
           onPressed: () async {
             try {
               String url = await OAuthClient.generateWebviewUrl(create: true);
-              _launch(url);
+              launch(url);
             } catch (e) {}
           },
           child: Text(I18n.of(context).dont_have_account),
@@ -76,7 +33,7 @@ class _FluentLoginPageState extends State<FluentLoginPage> {
             onPressed: () async {
               try {
                 String url = await OAuthClient.generateWebviewUrl();
-                _launch(url);
+                launch(url);
               } catch (e) {}
             }),
       ],
@@ -148,11 +105,13 @@ class _FluentLoginPageState extends State<FluentLoginPage> {
     );
   }
 
-  _launch(url) async {
+  @override
+  launch(url) async {
     try {
-      await launch(url);
+      await url_launcher.launch(url);
     } catch (e) {
       BotToast.showText(text: e.toString());
+      super.launch(url);
     }
   }
 }
